[Enhancement] [Counter] [Frequency is now 32 instead of 16 bit] [Cleanup] [Inputs-HAL] [getDiginFreq removed from HAL]
 - CounterI_c::frequency will now return 32bit.
 - getDiginFreq is only an internal function and shouldn't have been part of the HAL.

Done by M. Wodok
Postreview by M. Schmidt

+++ library/xgpl_src/IsoAgLib/hal/pc/system/system_target_extensions.h
@@ -43,7 +43,7 @@ public:
   virtual void init_analogin( uint8_t /*bNumber*/, uint8_t /*bType*/ ) {}
   virtual int16_t getDiginOnoff( uint8_t /*bInputNumber*/ ) { return false; }
   virtual void getDiginPeriod( uint8_t /*bInput*/, uint16_t * /*pwPeriod*/, uint16_t * /*pwImpulse*/ ) {}
-  virtual void getDiginFreq( uint8_t /*bInput*/, uint16_t * /*pwFrequency*/ ) {}
+  virtual void getDiginFreq( uint8_t /*bInput*/, uint32_t * /*pwFrequency*/ ) {}
   virtual int16_t getAdc( uint8_t /*bKanalnummer*/ ) { return 0; }
   virtual uint32_t getCounter( uint8_t /*bInput*/ ) { return 0; }
   virtual int16_t resetCounter( uint8_t /*bInput*/ ) { return 0; }
