+++ library/xgpl_src/supplementary_driver/hal/pc/inputs/target_extension_inputs_simulating.cpp
@@ -245,7 +245,7 @@ getDiginPeriod(uint8_t bInput, uint16_t *pwPeriod, uint16_t *pwImpulse)
 
 
 int16_t
-getDiginFreq (uint8_t bInputNumber, uint16_t *pwFrequency)
+getDiginFreq (uint8_t bInputNumber, uint32_t *pwFrequency)
 {
   if ( ! sensorDigitalInputOpen[bInputNumber] ) return HAL_CONFIG_ERR;
   checkForNewSimulatedDigitalValues (bInputNumber);
@@ -529,7 +529,7 @@ uint16_t getCounterPeriod(uint8_t ab_channel)
   @return frequency calculated from time between last two signals
           or 0 if time is longer than initially given timebase [100mHz]
 */
-uint16_t getCounterFrequency(uint8_t ab_channel)
+uint32_t getCounterFrequency(uint8_t ab_channel)
 {
   uint16_t ui16_timebase, ui16_result = 0;
   uint16_t ui16_lastSignalAge = 0;
